+++ grid2op/Observation/baseObservation.py
@@ -487,6 +487,16 @@ class BaseObservation(GridObjects):
         # gen up / down
         "gen_margin_up",
         "gen_margin_down",
+        # slack (>= 1.11.0)
+        "gen_p_slack",
+        # detachment (>= 1.11.0)
+        "load_detached",
+        "gen_detached",
+        "storage_detached",
+        "load_p_detached",
+        "load_q_detached",
+        "gen_p_detached",
+        "storage_p_detached",
     ]
 
     attr_list_vect = None
@@ -620,6 +630,18 @@ class BaseObservation(GridObjects):
         self.max_step = dt_int(np.iinfo(dt_int).max)
         self.delta_time = dt_float(5.0)
 
+        # slack (1.11.0)
+        self.gen_p_slack = np.empty(shape=cls.n_gen, dtype=dt_float)
+        
+        # detachment (>= 1.11.0)
+        self.load_detached = np.empty(shape=cls.n_load, dtype=dt_bool)
+        self.gen_detached = np.empty(shape=cls.n_gen, dtype=dt_bool)
+        self.storage_detached = np.empty(shape=cls.n_storage, dtype=dt_bool)
+        self.load_p_detached = np.empty(shape=cls.n_load, dtype=dt_float)
+        self.load_q_detached = np.empty(shape=cls.n_load, dtype=dt_float)
+        self.gen_p_detached = np.empty(shape=cls.n_gen, dtype=dt_float)
+        self.storage_p_detached = np.empty(shape=cls.n_storage, dtype=dt_float)
+        
     def _aux_copy(self, other : Self) -> None:
         attr_simple = [
             "max_step",
@@ -689,6 +711,16 @@ class BaseObservation(GridObjects):
             "gen_margin_up",
             "gen_margin_down",
             "curtailment_limit_effective",
+            # slack (>= 1.11.0)
+            "gen_p_slack",
+            # detachment (>= 1.11.0)
+            "load_detached",
+            "gen_detached",
+            "storage_detached",
+            "load_p_detached",
+            "load_q_detached",
+            "gen_p_detached",
+            "storage_p_detached",
         ]
 
         if type(self).shunts_data_available:
@@ -828,6 +860,9 @@ class BaseObservation(GridObjects):
                 - "theta" (optional) the voltage angle (in degree) of the bus to which the load is connected
                 - "bus" on which bus the load is connected in the substation
                 - "sub_id" the id of the substation to which the load is connected
+                - "detached" (>= 1.11.0) whether this load is detached from the grid
+                - "p_detached" (>= 1.11.0) amount of MW detached from the grid cause by the detachment of this load
+                - "q_detached" (>= 1.11.0) amount of MVAr detached from the grid cause by the detachment of this load
 
             - if a generator is inspected, then the keys are:
 
@@ -840,6 +875,18 @@ class BaseObservation(GridObjects):
                 - "actual_dispatch" the actual dispatch implemented for this generator
                 - "target_dispatch" the target dispatch (cumulation of all previously asked dispatch by the agent)
                   for this generator
+                - "curtailment": the curtailment applied on this generator (0. for non renewable generator)
+                - "curtailment_limit": the curtailment limit as given by the agent
+                - "curtailment_limit_effective": the effective curtailment limit
+                - "p_before_curtail": the active production (in MW) before any curtailment is applied
+                  this should be 0. for non renewable generator
+                - "margin_up": by how much this generateur can see its production increase between this step 
+                  and the next (in MW). It's 0. for renewable generators.
+                - "margin_down": by how much this generateur can see its production decrease between this step 
+                  and the next (in MW). It's 0. for renewable generators.
+                - "p_slack" (>= 1.11.0) slack participation (in MW) for this generator
+                - "detached" (>= 1.11.0) whether this generator is detached from the grid
+                - "p_detached" (>= 1.11.0) amount of MW detached from the grid cause by the detachment of this generator
 
             - if a powerline is inspected then the keys are "origin" and "extremity" each being dictionary with keys:
 
@@ -867,6 +914,9 @@ class BaseObservation(GridObjects):
                 - "storage_theta": (optional) the voltage angle of the bus at which the storage unit is connected
                 - "bus": the bus (1 or 2) to which the storage unit is connected
                 - "sub_id" : the id of the substation to which the sotrage unit is connected
+                - "detached" (>= 1.11.0) whether this storage is detached from the grid
+                - "p_detached" (>= 1.11.0) amount of MW detached from the grid cause by the detachment of this storage
+
 
             - if a substation is inspected, it returns the topology to this substation in a dictionary with keys:
 
@@ -926,6 +976,9 @@ class BaseObservation(GridObjects):
                 "v": self.load_v[load_id],
                 "bus": self.topo_vect[self.load_pos_topo_vect[load_id]],
                 "sub_id": cls.load_to_subid[load_id],
+                "detached": self.load_detached[load_id],
+                "p_detached": self.load_p_detached[load_id],
+                "q_detached": self.load_q_detached[load_id],
             }
             if self.support_theta:
                 res["theta"] = self.load_theta[load_id]
@@ -959,6 +1012,9 @@ class BaseObservation(GridObjects):
                 "p_before_curtail": self.gen_p_before_curtail[gen_id],
                 "margin_up": self.gen_margin_up[gen_id],
                 "margin_down": self.gen_margin_down[gen_id],
+                "gen_p_slack": self.gen_p_slack[gen_id],
+                "detached": self.gen_detached[gen_id],
+                "p_detached": self.gen_p_detached[gen_id],
             }
             if self.support_theta:
                 res["theta"] = self.gen_theta[gen_id]
@@ -1026,6 +1082,8 @@ class BaseObservation(GridObjects):
             res["storage_power_target"] = self.storage_power_target[storage_id]
             res["bus"] = self.topo_vect[cls.storage_pos_topo_vect[storage_id]]
             res["sub_id"] = cls.storage_to_subid[storage_id]
+            res["detached"] = self.storage_detached[storage_id],
+            res["p_detached"] = self.storage_p_detached[storage_id],
             if self.support_theta:
                 res["theta"] = self.storage_theta[storage_id]
         else:
@@ -1171,6 +1229,28 @@ class BaseObservation(GridObjects):
             except ValueError as exc_:
                 # this attribute was not there in the first place
                 pass 
+            
+    @classmethod
+    def _aux_process_grid2op_compat_1_11_0(cls):
+        cls.attr_list_vect = copy.deepcopy(cls.attr_list_vect)
+
+        for el in [
+            # slack (>= 1.11.0)
+            "gen_p_slack",
+            # detachment (>= 1.11.0)
+            "load_detached",
+            "gen_detached",
+            "storage_detached",
+            "load_p_detached",
+            "load_q_detached",
+            "gen_p_detached",
+            "storage_p_detached",
+        ]:
+            try:
+                cls.attr_list_vect.remove(el)
+            except ValueError as exc_:
+                # this attribute was not there in the first place
+                pass 
         
     @classmethod
     def process_grid2op_compat(cls) -> None:
@@ -1200,6 +1280,10 @@ class BaseObservation(GridObjects):
         if glop_ver < version.parse("1.9.1"):
             # alert attributes have been added in 1.9.1
             cls._aux_process_grid2op_compat_191()
+            
+        if glop_ver < version.parse("1.11.0"):
+            # alert attributes have been added in 1.9.1
+            cls._aux_process_grid2op_compat_1_11_0()
             
         cls.attr_list_set = copy.deepcopy(cls.attr_list_set)
         cls.attr_list_set = set(cls.attr_list_vect)
@@ -1326,6 +1410,18 @@ class BaseObservation(GridObjects):
         self.curtailment_limit_effective[:] = 0.
         self.curtailment[:] = 0. 
 
+        # slack (>= 1.11.0)
+        self.gen_p_slack[:] = 0.
+        
+        # detachment (>= 1.11.0)
+        self.load_detached[:] = True
+        self.gen_detached[:] = True
+        self.storage_detached[:] = True
+        self.load_p_detached[:] = 0.
+        self.load_q_detached[:] = 0.
+        self.gen_p_detached[:] = 0.
+        self.storage_p_detached[:] = 0.
+        
     def set_game_over(self,
                       env: Optional["grid2op.Environment.Environment"]=None) -> None:
         """
@@ -1463,6 +1559,18 @@ class BaseObservation(GridObjects):
         # was_alert_used_after_attack not updated here in this case
         # attack_under_alert not updated here in this case
 
+        # slack (>= 1.11.0)
+        self.gen_p_slack[:] = 0.
+        
+        # detachment (>= 1.11.0)
+        self.load_detached[:] = True
+        self.gen_detached[:] = True
+        self.storage_detached[:] = True
+        self.load_p_detached[:] = 0.
+        self.load_q_detached[:] = 0.
+        self.gen_p_detached[:] = 0.
+        self.storage_p_detached[:] = 0.
+        
     def __compare_stats(self, other: Self, name: str) -> bool:
         attr_me = getattr(self, name)
         attr_other = getattr(other, name)
@@ -2706,13 +2814,24 @@ class BaseObservation(GridObjects):
         return bus_ids
         
     def _aux_add_loads(self, graph, cls, first_id):
+        nodes_prop = [
+            ("detached", self.load_detached),
+            ("p_detached", self.load_p_detached),
+            ("q_detached", self.load_q_detached),
+        ]
         edges_prop=[
             ("p", self.load_p),
             ("q", self.load_q),
-            ("v", self.load_v)
+            ("v", self.load_v),
         ]
         if self.support_theta:
             edges_prop.append(("theta", self.load_theta))
+
+        # slack (>= 1.11.0)
+        self.gen_p_slack[:] = 0.
+        
+        if "load_detached" in self.attr_list_set:
+            edges_prop.append(("is_detached", self.load_detached))
         load_ids = self._aux_add_el_to_comp_graph(graph,
                                                   first_id,
                                                   cls.name_load,
@@ -2720,7 +2839,7 @@ class BaseObservation(GridObjects):
                                                   cls.n_load,
                                                   self.load_bus,
                                                   cls.load_to_subid,
-                                                  nodes_prop=None,
+                                                  nodes_prop=nodes_prop,
                                                   edges_prop=edges_prop)
         return load_ids
     
@@ -2732,7 +2851,10 @@ class BaseObservation(GridObjects):
                       ("curtailment", self.curtailment),
                       ("curtailment_limit", self.curtailment_limit),
                       ("gen_margin_up", self.gen_margin_up),
-                      ("gen_margin_down", self.gen_margin_down)
+                      ("gen_margin_down", self.gen_margin_down),
+                      ("p_slack", self.gen_p_slack),
+                      ("detached", self.gen_detached),
+                      ("p_detached", self.gen_p_detached),
                       ]  # todo class attributes gen_max_ramp_up etc.
         edges_prop=[
             ("p", - self.gen_p),
@@ -2754,7 +2876,10 @@ class BaseObservation(GridObjects):
     
     def _aux_add_storages(self, graph, cls, first_id):
         nodes_prop = [("storage_charge", self.storage_charge),
-                      ("storage_power_target", self.storage_power_target)]  
+                      ("storage_power_target", self.storage_power_target),
+                      ("detached", self.storage_detached),
+                      ("p_detached", self.storage_p_detached),]  
+        
         # TODO class attr in nodes_prop: storageEmax etc.
         edges_prop=[("p", self.storage_power)]
         if self.support_theta:
@@ -3766,6 +3891,8 @@ class BaseObservation(GridObjects):
             # current_step / max step
             self._dictionnarized["current_step"] = self.current_step
             self._dictionnarized["max_step"] = self.max_step
+            
+            # TODO shedding: add relevant attributes
 
         return self._dictionnarized
 
@@ -3965,7 +4092,7 @@ class BaseObservation(GridObjects):
         cls = type(self)
         cls_act = type(act)
         
-        act = copy.deepcopy(act)
+        act : BaseAction = copy.deepcopy(act)
         res = cls()
         res.set_game_over(env=None)
 
@@ -3977,8 +4104,13 @@ class BaseObservation(GridObjects):
             raise RuntimeError(
                 f"Impossible to add an ambiguous action to an observation. Your action was "
                 f'ambiguous because: "{except_tmp}"'
-            )
+            ) from except_tmp
 
+        if act._modif_detach_gen or act._modif_detach_load or act._modif_detach_storage:
+            raise NotImplementedError("This function is not yet implemented when some elements "
+                                      "are detached from the grid. Please write a feature request "
+                                      "if you are interested in this feature.")
+            
         # if a powerline has been reconnected without specific bus, i issue a warning
         if "set_line_status" in cls_act.authorized_keys:
             self._aux_add_act_set_line_status(cls, cls_act, act, res, issue_warn)
@@ -4262,6 +4394,18 @@ class BaseObservation(GridObjects):
             self.curtailment_limit_effective[:] = 1.0
 
         self.delta_time = dt_float(1.0 * env.delta_time_seconds / 60.0)
+
+        # slack (1.11.0)
+        self.gen_p_slack[:] = env._slack_gen_p
+        
+        # detachment (>= 1.11.0)
+        self.load_detached[:] = env._loads_detached
+        self.gen_detached[:] = env._gens_detached
+        self.storage_detached[:] = env._storages_detached
+        self.load_p_detached[:] = env._load_p_detached
+        self.load_q_detached[:] = env._load_q_detached
+        self.gen_p_detached[:] = env._gen_p_detached
+        self.storage_p_detached[:] = env._storage_p_detached
         
         # handles forecasts here
         self._update_forecast(env, with_forecast)